Stub out UI for simpler workflow
I was irritated by how I had to load javascript urls one-at-a-time.
While working to get the most recent version of the react-datepicker's
dependencies to work together, I had to fiddle with this many times
and thought it'd be easier to just allow specification of all the
desired files at once.

This branch sets up the start of that workflow, without hooking into
the actual functions that do the work yet.